--- a/Cargo.toml
+++ b/Cargo.toml
@@ -50,6 +50,7 @@ unicode-segmentation = "1.8.0"
 levenshtein = "1.0.5"
 syslog = "6.0.0"
 defaults = "0.2.0"
+freedesktop-icon-lookup = "0.1.0"
 
 [profile.release]
 lto = true