resolves #13

resolves #76