@l4l 2022-10-11
- Configure time: 199ms -> 28ms
- First draw time: 34ms -> 63ms

i.e startup time ~x2.5 faster :tada: