@mrpandey 2022-10-31
After updating to these changes (release 0.2.0), the icons of most apps defaulted to the fallback icon. Just had to comment out the following line in the config file to make it work: `theme = "theme-name"`.