@l4l 2022-10-31
Yes, indeed now the default icon theme is `hicolor` as specified in [the spec](https://specifications.freedesktop.org/icon-theme-spec/icon-theme-spec-latest.html). And previously `/usr/share/icons/default/index.theme` was read to get one which seems to be about cursor theme. So formally it was an incorrect behavior before.

One need to specify theme now, usually `theme = "Adwaita"`. Clarified a= noted in the [Config wiki](https://github.com/l4l/yofi/wiki/Configuration). Thanks for this one